+++ Jolt/Core/LinearCurve.cpp
@@ -21,12 +21,12 @@ JPH_IMPLEMENT_SERIALIZABLE_NON_VIRTUAL(LinearCurve)
 	JPH_ADD_ATTRIBUTE(LinearCurve, mPoints)
 }
 
-float LinearCurve::GetValue(float inX) const
+decimal LinearCurve::GetValue(decimal inX) const
 {
 	if (mPoints.empty())
 		return 0.0f;
 
-	Points::const_iterator i2 = lower_bound(mPoints.begin(), mPoints.end(), inX, [](const Point &inPoint, float inValue) { return inPoint.mX < inValue; });
+	Points::const_iterator i2 = lower_bound(mPoints.begin(), mPoints.end(), inX, [](const Point &inPoint, decimal inValue) { return inPoint.mX < inValue; });
 
 	if (i2 == mPoints.begin())
 		return mPoints.front().mY;
