+++ Jolt/Physics/Collision/CollidePointResult.h
@@ -15,7 +15,7 @@ public:
 	JPH_OVERRIDE_NEW_DELETE
 
 	/// Function required by the CollisionCollector. A smaller fraction is considered to be a 'better hit'. For point queries there is no sensible return value.
-	inline float	GetEarlyOutFraction() const			{ return 0.0f; }
+	inline decimal	GetEarlyOutFraction() const			{ return decimal(0.0f); }
 
 	BodyID			mBodyID;							///< Body that was hit
 	SubShapeID		mSubShapeID2;						///< Sub shape ID of shape that we collided against
